Refactor deprecation helpers out into decorators
